Move question group middleware to own module
